Add UnmarshalManyPayloadWithLinks function
This commit adds a new function UnmarshalManyPayloadWithLinks which is a
copy of UnmarshalManyPayload but it also returns a map of the pagination
links.

The code of the two functions should be maintained to be equivalent.

The purpose of the new function is to give access to the pagination
links which are included in the JSON API document. This function is the
reason for keeping this fork of google/jsonapi which provides a decent
solution for the pagination links issue described in:
https://github.com/nstratos/go-kitsu/issues/2